Abstracted Mission Class & Added Logic - Added logic to handle at the end of the mission - Moved all "abstract" logic to the Mission class, and left all mission-specific code in ISRMission
Reorganized Project structure to separate DataStructures with Logic 
classes.

Added more getter logic for the ListDict to allow for custom get methods 
per key.
